@@ -1,36 +1,21 @@
-import { Alchemy, Network } from 'alchemy-sdk';
-import { useEffect, useState } from 'react';
+import { useEffect, useState } from "react";
+import BlockCard from "./components/BlockCard/BlockCard";
+import alchemy from "./utils/alchemy/alchemy-configuration";
 
-import './App.css';
-
-// Refer to the README doc for more information about using API
-// keys in client-side code. You should never do this in production
-// level code.
-const settings = {
-  apiKey: process.env.REACT_APP_ALCHEMY_API_KEY,
-  network: Network.ETH_MAINNET,
-};
-
-
-// In this week's lessons we used ethers.js. Here we are using the
-// Alchemy SDK is an umbrella library with several different packages.
-//
-// You can read more about the packages here:
-//   https://docs.alchemy.com/reference/alchemy-sdk-api-surface-overview#api-surface
-const alchemy = new Alchemy(settings);
+import "./App.css";
 
 function App() {
-  const [blockNumber, setBlockNumber] = useState();
+  const [block, setBlock] = useState();
 
   useEffect(() => {
-    async function getBlockNumber() {
-      setBlockNumber(await alchemy.core.getBlockNumber());
+    async function getBlock() {
+      setBlock(await alchemy.core.getBlock());
     }
 
-    getBlockNumber();
+    getBlock();
   });
 
-  return <div className="App">Block Number: {blockNumber}</div>;
+  return <div className="App">{block && <BlockCard block={block} />}</div>;
 }
 
 export default App;
